ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3

pytesseract==0.3.13
Pillow==11.0.0
PyMuPDF==1.24.14

ultralytics==8.3.27
opencv-python==4.10.0.84

llama-cpp-python==0.3.7
huggingface-hub==0.26.2
